FIX changed datatype for year in Date.cs from short to long

+++ Date.cs
@@ -11,14 +11,14 @@ namespace Calendar
         List<DateEntry> entries;
         byte day;
         byte month;
-        short year;
+        long year;
 
         //Constructors
 
         public Date(DateTime dt) //should be the first second of the day, as in 00:00:00
         {
             List<string> list = dt.GetDateTimeList(3);
-            short.TryParse(list[0], out year);
+            long.TryParse(list[0], out year);
             byte.TryParse(list[1], out month);
             byte.TryParse(list[2], out day);
             entries = null;
@@ -30,7 +30,7 @@ namespace Calendar
 
         public byte Day { get => day; set => day = value; }
         public byte Month { get => month; set => month = value; }
-        public short Year { get => year; set => year = value; }
+        public long Year { get => year; set => year = value; }
         internal List<DateEntry> Entries { get => entries; set => entries = value; }
     }
 }
